Fix Typo in Instructions, Correct run_map_reduce
Remove flatten from the mapping step.

More concretely, change

`pairs = flatten(list(map(map_f, mylist)))`
to
`pairs = list(map(map_f, mylist))`

In this problem, `mylist` is a single list of values. If we were
performing map-reduce on a list of lists, then we would
need to flatten, but here we will end up with a single list
of tuples which do not need to be flattened.